Multiple changes to backend
 * Renamed ArticleImage to Image.
 * Established a many-to-many relationship between images and articles.
 * Implemented a separate API endpoint to link images to articles.
 * Added policies for both images and articles.
 * Updated policies related to articles.
 * Modified factory to accept user and post when creating postable content.

Faker Seeders:

 * Introduced faked method to generate fake articles.
 * Removed ArticleImageFactory.
 * Set files as public by default.
 * Associated fileable models in FileFactory via the fileable method.
 * Linked uploadFile method in FileFactory to an UploadedFile instance.
 * Associated fake images in ImageFactory using fakedImage method.
 * Specified file information using the withFile method in ImageFactory.
 * Associated files, users, and articles with images during seeding.
 * Assigned a random main image for each article.
 * Established the main image association in ArticleImageSeeder.
 * Allowed specification of seeding user for articles.
 * Associated between 1 and 3 tags with each article.
 * Linked article users to images.
 * Provided an option to specify a pool of users for seeding comments.
 * Enabled specification of the number of comments to seed.
 * Implemented seeding for multiple users.
 * UserSeeder now triggers ImageSeeder and CommentSeeder.

Initial Seeders:

 * Created posts for articles.
 * Associated users with articles.
 * Resolved issues with revisions not being linked to articles.

+++ app/Http/Controllers/Api/Blog/ImageController.php
@@ -3,8 +3,7 @@
 namespace App\Http\Controllers\Api\Blog;
 
 use App\Http\Controllers\Controller;
-use App\Models\Article;
-use App\Models\ArticleImage;
+use App\Models\Image;
 use App\Models\File;
 use Illuminate\Http\Request;
 
@@ -13,74 +12,62 @@ class ImageController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(Article $article)
+    public function index(Request $request)
     {
-        return $article->images;
+        $this->authorize('index', Image::class);
+
+        return $request->user()->hasRoles(['admin']) ? Image::all() : Image::owned($request->user())->get();
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request, Article $article)
+    public function store(Request $request)
     {
+        $this->authorize('create', Image::class);
+
         $request->validate([
             'image' => 'required|image',
             'description' => 'nullable|string|max:255',
         ]);
 
-        $articleImage = new ArticleImage([
+        $image = new Image([
             'description' => isset($request->description) ? $request->description : null,
         ]);
 
         $path = $request->file('image')->store('images');
         $file = File::createFromFilePath($path, null, true);
 
-        $article->images()->save($articleImage)->file()->save($file);
+        $file->user()->associate($request->user());
 
-        return $articleImage->load('file');
+        $image->save();
+
+        $image->file()->save($file);
+
+        return $image->load('file');
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Article $article, ArticleImage $image)
+    public function show(Image $image)
     {
+        $this->authorize('view', $image);
+
         return $image;
-    }
-
-    /**
-     * Sets main image for article.
-     */
-    public function mainImage(Request $request, Article $article, ArticleImage $image)
-    {
-        $article->mainImage()->associate($image);
-        $article->save();
-
-        return $article;
-    }
-
-    /**
-     * Removes main image for article.
-     */
-    public function destroyMainImage(Request $request, Article $article)
-    {
-        $article->mainImage()->dissociate();
-        $article->save();
-
-        return $article;
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Article $article, ArticleImage $image)
+    public function destroy(Image $image)
     {
-        // TODO: Don't allow if set as main image.
+        $this->authorize('delete', $image);
 
         $image->delete();
 
         return [
-            'success' => __('Article image ":name" was removed.', ['name' => $image->file->name]),
+            'success' => __('Image ":name" was removed.', ['name' => $image->file->name]),
         ];
     }
 }
